Move strategies to separate file
